Odyssey FFDC: Connect to FSP
This commit connects the Odyssey FIFO FFDC dump chipop to FSP command
line. A user can use a putcfam command to execute a custom istep on FSP
in between Hostboot isteps to trigger an Odyssey FFDC dump. The command is
executed like so (from FSP command line):

putcfam pu 283b D500FF07

This command requests a custom istep (255.7) to be executed in HB, which
performs an Odyssey FFDC dump. The output will be logged to error
log(s). Note that this command needs to be executed at a "stable" point
in between isteps to ensure that no other chipops are going to Odyssey
chips. The command will request the FFDC dump from all functional
Odyssey chips in the system.

Change-Id: I47e0531e20e905cf4e1c684a8e33f83cc0e36c38
JIRA: PFHB-254
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/158696
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>