Only use 31 bits for OCMB MMIO address comparisons
The OpenCAPI/OMI interface allows for 35 bits of address space,
however we only configure the OCMBs to use 2GB of that space.
That means that when we compare addresses between internal
registers on the OCMB (e.g. MMIOERR) with the physical address
we use at the system level, we should only be looking at the
bottom 31 bits.

Change-Id: If8df2b4b73dde3dc967d3c960d21b495ee345e2e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/159078
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>