Prime Cache Bugs
In testing of prime caching, I missed the scenario when HB_FAST_PRIME is
not set and the cache exists for the local levels of the submodules. In
this case, we still want to force-rebuild the images and not rely on the
cache. This commit fixes some issues with this path.

Change-Id: I83897c547f2e75a30e416b607ee6535e3978ac41
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/159185
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>