Check for Odyssey Async FFDC
This commit implements logic to check for Async FFDC from
Odyssey after Odyssey boots. If SBE indicated that there is
async FFDC present, the logic sends the getFFDC chip op to
SBE and commits the resulting error logs. If the code level on
the given SBE is current, the error logs are committed with the
reported severity. If the code level is not current, the severity
of the errors is reduced to informational, since there is a chance
that the code update would fix the issues reported, and we don't
want to have visible error logs with callouts.

Change-Id: I95234172906bc0172905f0b03951acf2c42ce3da
JIRA: PFHB-246
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/159186
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>