SBE FFDC Parser: Update sort criteria
This commit updates the sort criteria for odyssey ffdc packages to
further sort packages first by SLID then by HWP < PLAT < RC=0 in an
effort to reduce the number of logs generated on a failure.

The logic to generate the logs based on this sorting criteria will be
added in a subsequent commit.

Change-Id: I2c6f872af57735669017aad46c7b92b9a7077d8a
JIRA: PFHB-601
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/160309
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>