Re-enable MSR[ME] after taking a machine-check instruction if necessary
BACKGROUND
----------

A machine-check is an error detected in peripheral hardware, e.g. a
memory chip. This can happen when, for example, an invalid SCOM
address is accessed. When a machine-check occurs, a machine-check
interrupt is signalled to the thread that requested the operation that
failed.

When a machine-check interrupt happens, the hardware clears the
"machine-check interrupt enable" bit (ME) in the machine state
register (MSR). If MSR[ME] is already 0 when a machine-check exception
occurs, the system checkstops. This effectively detects and halts the
situation when a machine-check exception occurs while handling another
machine-check exception; software can't recover from nested errors, so
the system checkstops in this case.

Hostboot sets MSR[ME] to 1 early in the boot so that we can handle
machine checks in software and, for example, communicate to the caller
of an MMIO operation that their requested read or write failed.

THIS COMMIT
-----------

There was a bug previously, where Hostboot would never re-enable
machine-check exceptions by setting MSR[ME] to 1 again, after hardware
had cleared it when a machine-check interrupt occurred.

This commit fixes the issue by adding a parameter to the function that
sets MSR after taking these interrupts, namely kernel_dispatch_task,
that allows callers to request that it set MSR[ME] to 1. This
parameter is only set to 1 in the machine-check interrupt path,
because if we're in that path it means that Hostboot has already set
MSR[ME]=1 (and Hostboot does not ever want it to stay clear after
handling a machine-check interrupt). The parameter is 0 in all other
paths, which will cause kernel_dispatch_task to leave MSR[ME]
unmodified (and MSR[ME] is preserved by all other interrupt types).

Change-Id: I9f6c985e7e71dddec3ffdfaaa2c29a217558c484
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/159685
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>