Create SBE FFDC parser log generator
This commit adds support for a log generator for Odyssey SBE packages.
The logic will parse through the list of FFDC packages and group the
data by SBE log Id and return a list of errors back to the caller for
them to handle.

This is new for Odyssey where there can be a number of unrelated errors
returned from a single operation to the SBE. For P10, the commit just
moves some additional info off of the chip-op failure log and onto the
log where the original error context is from.

Change-Id: I7731268b664249340450e34477ab0260c05995a1
JIRA: PFHB-601
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/160310
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>