Change waitUpFifoReady exit trace back to TRACU
Seeing a watchdog timeout in istep 12.2 omi_attr_update. The exit
trace for waitUpFifoReady was recently changed from TRACU to TRACF,
likely for testing reasons. This resulted in an excessive number
of these traces in the hb output when traces are enabled. Suspect
these extra traces are causing extra traffic and ultimately the
watchdog timeout.  This commit changes the trace back to TRACU.

Change-Id: Ia1b2ebf6b70741459196e67a66dc08d33b4d561f
STGD:582966
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/159781
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>