Enable Trace Entry resiliency
If an unrecognized entry type is detected in the parsing of the trace
data, output an information warning to identify that there are issues
with the binary data, but continue to process the remaining valid
entries.

This problem surfaced when a Hostboot Runtime RSVD_MEM_TRACE was
improperly formatted (see rtPost->callCommitRsvdTraceBufErrl for
how a runtime crash preserves the trace buffer for problem
analysis on a subsequent boot), the next IPL will log the crash
RSVD_MEM_TRACE buffer).

Change-Id: I3ce8d3e5b2a8c2248f7676763a484c707fd02d15
STGD:  585828
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/160565
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>