Stop forcing rhel7 for startsimics
Simics moved to rhel8 a short while ago. When hostboot developers attempt to
start simics the hb util will try to match the rhel level of their
machine. This will cause simics to fail to start with a complaint about
GLIBCXX_3.4.21 not being found if the dev is starting from a rhel7
machine.

To prevent this, just run simics on a rhel8 machine regardless of the
dev's machine

Change-Id: Iad71a1520ec9823fca23f8696e6750fe232ca81d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/160571
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>