Odyssey Secureboot: SPPE Measurement Reg Verification
This commit implements the second part of the Odyssey secureboot
verification - the read and verification of the Odyssey measurement
registers. These registers (SCOM 0x501A0-0x501AB for bootloader and
SCOM 0x501AC-0x501B7 for runtime) are written with the hash of the
measurement of the current code that's running on SPPE. This hash
mush match the measured.hash from the OCMBFW PNOR partition for both
bootloader and runtime images. Note that this hash will be calculated
correctly on SPPE only if bits 16-22 of ATTR_OCMB_BOOT_FLAGS are set.

Change-Id: Iaa4ceacc639bca0401f04ae81a8cadba17c29070
JIRA: PFHB-455
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/160366
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Chris Engel <cjengel@us.ibm.com>