Use reasonable timeout for exp_omi_setup
The current polling logic assumes that the i2c operation itself
takes zero time.  Therefore it will take 100,000 loops to fail.
In reality the operation takes nonzero time so that many loops
ends up taking essentially forever.

This change takes advantage of a recent change in the HWP that
allows the caller to control the overall total time spent.

Change-Id: I7c534d81205de26d27b543740cdbaff420db11a4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/161093
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>