Update patch-bmc script with new Usage in help text
This commit adds Usage(3) to the patch-bmc help text which explains
how to use the script on a directory that contains all of the necessary
data to patch a system.  In other words, this new Usage(3) shows how to
use the script after the --cponly option (aka Usage(2)) is used.

Change-Id: I3b620e99c2b16d3cd5595a0237284bcd3ee14390
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/160995
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Ilya Smirnov <ismirno@us.ibm.com>