Add support for aggregating ErrlEntry objects
This commit adds some functionality to the ErrlEntry class to allow it
to "aggregate" with other logs, such that the group of logs behaves as
if it were a single log; when the toplevel log is committed or
deleted, all logs in the aggregate are committed or deleted.

This allows code that was written to handle only a single error log
to transparently handle more than one.

Certain member functions, like setSev, propagate to all logs in an
aggregate.

Note that this abstraction is only visible to code *before* the log
reaches the Error log manager via errlCommit(). At the time the log
is committed, the aggregate is broken apart and each log is committed
separately; the ErrlManager internals are completely unaware of the
aggregate functionality. Because of this, many of the private
functions of the ErrlEntry class are not aggregate-aware, because
they are only used after any aggregate has been broken apart and the
log is being considered on its own again.

Change-Id: Ie3e21fba1058be9f927b1eba05b895381ab7fca7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/161015
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>