Create DYNAMIC_IO_DRAWER_ATTACHMENT Attribute
This commits adds the DYNAMIC_IO_DRAWER_ATTACHMENT attribute to track settings
on a per-node level for TCE Table Pre-Allocation for Dynamic IO Drawer
Attachment support.  It is only needed for OpenPower PHYP configurations,
which is why it was put into the "oppowervm" attribute_types_ and target_types_
XML files.

Change-Id: I8333e08c24815341de531137a9492f71f72f6208
JIRA: PFHB-589
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/161178
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>