PM HWP: make PPC405 cache UE FIR report as a recoverable attn to PRD
This bug carried over from P9 whereby a PPC405 cache UE was being
reported to the 405 itself.  This has an obvious hazard of possibly
allowing the OCC FW to actually report the error.  This commit changes
the OCC LFIR setting for this bit to RECV_ATTN to surface this to PRD.

Change-Id: I9dd44a351f8c4cbe5d71b1f20397296e270eff9f
STGD: 588015
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb/+/163204
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/163394
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>