Run copy of EECACHE to alt pnor in the background
Since the alt pnor must be done via XScom LPC access, it
takes 40-45 seconds.  Create a background task to perform
this in parallel to the mainline IPL flow.

This task is part of the pnor lib (not istep as it gets
unloaded).  Register with the shutdown event queue to
allow this task to finish prior to a controlled shutdown.

Change-Id: Ibd6b53d4395e75e66365a50fefcc80bc9067f911
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/181265
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>