Odyssey code update: Only invoke the FSM for async FFDC if the boot failed
This commit makes a change to Istep 11 (ocmb_check_for_ready) so that
the Odyssey code update FSM is only invoked with the "async FFDC"
event if the boot failed (i.e. if an error was returned by a boot
HWP). If the boot didn't fail but we got async FFDC packets, we just
commit the async FFDC errors as-is and continue the boot
(since the presence of async FFDC alone does not indicate that the
part will not be able to function).

Change-Id: If3505d8a9b0abb5cf48db009bda715842ccabec4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/162984
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>