Add OCMB Chip ID to HDAT
PHYP needs a way to differentiate between Explorer and Odyssey
OCMB chips.  A new field containing the chipid (IDEC) value is
being added to the end of the "Address range array for MMIO
regions for OCMB access by HBRT".
  60D2=Explorer
  60C0=Odyssey
Note that this new field requires an update the the version
of the MSAREA from 0x50 to 0x51.

Change-Id: Ie4c78cbe950330e549416ec2aa094372d28a1b5d
JIRA: PFHB-689
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/162996
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Jayashankar Padath <jayashankar.padath@in.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>