PRD: update handling of OCC SRAM ECC errors
OCC_FIR[45] PPC405 cache CE is currently set to threshold on first
occurrence. Instead, it should be informational only. OCC_FIR[44]
PCC405 cache UE is currently set to be handled by the OCC/PM firmware.
However, if that firmware is running in the SRAM with a UE, it may crash
instead of report the error. Therefore, PRD will handle the attention
instead and make a callout on first occurrence.

Change-Id: Ifcb5ff93c885c4478adc9ffe5d2e986c195959a0
STGD: 588015
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/prd/+/163090
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/prd/+/163212
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/163220
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>