PRD: set L2FIR (RC) load/store received as SUE source
L2FIR[17:18] (RC) load/store received are both set to checkstop root
cause, but the UERE flag is not set. So if paired with an ME=0 checkstop
(which means hardware saw an SUE), PRD will callout the right part, but
the EX will not be guarded because the UERE flag is not set.

STGD: 590738
Change-Id: I811650814eb9f570a83a93ca48abec72643d04f3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/prd/+/163190
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/163242
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>