Odyssey SBE main_sbe_handler Bad Path
Provide the recovery mechanisms when the Odyssey SBE encounters
any errors during the hreset processing.

If the Odyssey SBE fails to properly return to a runtime state,
perform a side switch and attempt to boot to runtime.

If all mechanisms fail to hreset successfully, collect a dump
and mark the OCMB with a state of DEAD (so that subsequent
hreset requests will be basically ignored).

Change-Id: I31e3fcfe164669556a3e95c924e83f5495b7ecae
JIRA:  PFHB-542
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/159925
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>