Handle missing memory mirroring BIOS attribute
Some systems do not support memory mirroring(Bonnell).
Mirroring is enabled via the BIOS Attribute hb_memory_mirror_mode.
This is controlled via various BMC UI methods.  Hostboot then
consumes the value, uses it to configure the system, then tells
PHYP if mirroring is enabled or not. This commit sets the
HB memory mirror attribute to 0 if the BIOS attribute does not
exist. It also removes the error log commit as this is expected
behavior.

Change-Id: I99f77fe895f72d9ad74250233ba6368125de7259
JIRA:PFHB-611
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/163596
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>