Handle Odyssey targets in FSP SBE HRESET request
BACKGROUND
==========

The Odyssey SBE firmware is responsible for obtaining DIMM temperature
information and providing them to the SP in SCOM registers, which the
SP reads.

When the SP detects that the temperature information is out of date,
it means either:

1. The SPPE crashed and is not reading the temperature sensors, or
2. there were errors reading the temperature sensors, or
3. there's some code bug.

In all cases, we want to (A) get any FFDC from the SBE so that we have
data on the failure if any data exists, and (B) perform an HRESET on
the SBE.

(If the HRESET fails, we will dump the SBE, which will be implemented
in another story.)

IMPLEMENTATION
==============

There exists an FSP mailbox command called MSG_SBE_ERROR,
parameterized on a (processor) HUID and an error log ID, which the FSP
sends to HBRT in order to invoke the SBE Retry Handler on a given
processor.

This commit expands MSG_SBE_ERROR to support Odyssey OCMBs as well.

It also adds an error class, called SBEIO_ERROR_TYPE_HRESET_PERFORMED,
to indicate that an error is associated with an HRESET, so that when
checking for FFDC, if the chip operation times out and triggers an
HRESET, we don't do more than one.

JIRA: PFHB-561
Change-Id: Id457fab3ac7ce7d01abf46c2338aa3f13ecb0a0f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/163823
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>