More efficient pnor write in default guard record case
The code as previously written was flushing the entire pnor
partition out after updating each record.  This was very slow,
especially in simics where it added close to a minute.  A more
efficient implementation is to do all the writes and then force
a single flush of the entire partition.

Change-Id: I607455e4de895a4c8b98ccbdd2cf38215b2ca81f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/163675
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>