Do not parse FFDC from getFFDC chip-op in SBE FIFO Device Driver
There was an issue in the FIFO device driver where a caller would
request the ASYNC FFDC via the get FFDC chip-op and the device driver
would blindly parse this data and return it back as errors. After which,
callers may act on these errors inappropriately.

Now calls to the get FFDC chip-op will have the SBE FIFO DD ignore FFDC
returned back as part of the chip-op unless the chip-op itself fails.

Change-Id: Ib4fb167b5ffa99c700b5caaad44e6913b87cffd7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/164068
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>