Increase pagemgr Timeout
We've encoutered an intermittent situation where PRD attempts to
allocate more memory pages than the pagemgr has free. This happens in
istep 14.1 when we don't have full memory yet. To remedy this, the
least intrusive fix is to increase the number of iterations the
pagemgr performs looking for pages before crashing with OOM. This
will give the threads that execute in 14.1 more of a chance to
finish and release memory to the pool. Traces show that the current
number of iterations take rougly .4 seconds, so it's safe to increase
the timeout tenfold.

Change-Id: I3d3eb2405cb87d75a749918ceb7b8beffad97047
STGD: 409634
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/164415
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>