Add MAX_*_PER_OCMB to convertSbeTargInstanceToFapiPos
There was a bug where the number of PERVASIVE targets on Odyssey OCMBs
was not present in the code which converts SBE target instances to
FAPI_POS.

This commit addresses that issue so the code can convert those values to
be used for callout/deconfig/guard later.

Additionally, adds MAX_DIMM_PER_OCMB to resolve a similar issue.

Change-Id: I320265b5e247f68d44b070081e48c11a262506af
JIRA: PFHB-551
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/164450
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Ilya Smirnov <ismirno@us.ibm.com>