Fix getTarget instance parameter
There was an issue where the getTarget parameter for FAPI_POS instance
was uint8_t and Odyssey PERV targets have 0x8000 added onto their
FAPI_POS to differentiate from P10 PERV targets. As a result, a
narrowing conversion would occur when Odyssey PERV instance values are
passed to the function and the added 0x8000 would be dropped making it
impossible to distingiush P10 and Odyssey targets. This leads to
incorrect callouts in some error paths.

This commit updates the getTarget parameter to be wide enough to fit the
added 0x8000.

Change-Id: Ia3f4f00e2a25cbe012247090cdab6d96a07343f4
JIRA: PFHB-551
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/164505
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>