Odyssey Secureboot: Secure Reg Checking Changes
In manufacturing (mfg) mode, the OCBM_BOOT_FLAGS attribute is not
being set to include the secureboot settings for Odyssey because
in mfg mode, it is expected that the security settings are already
fused into the DIMMs. Before this change, Odyssey secureboot
verification logic relied on OCMB_BOOT_FLAGS attribute to compare
the expected security settings to actual settings reported by
Odyssey. This change removes the logic to read the attribute
and compare the contents of the Ody secure reg to the bits in
the attribute, and instead compares the contents of the reg to
values we've agreed upon prior. This will allow the verification
of the secure register to pass in mfg mode.

Additionally, now the verification logic checks bit 9 of the
BL_SB_CONTROL_REG (SCOM 0x501B9) and compares it to production
mode setting on the system.

Change-Id: Ia42b877e393a85f19bd4c75d41a52952610281e3
STGD: 595891
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/164508
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Chris Engel <cjengel@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>