Added SBE_ATTRIBUTE_RC_DENIED_UPDATE for soft security check
SBE soft security functions are controlled by ATTR_SECURITY_LEVEL. A new tag
denyForSecurityUpdate has been added to this attribute in the attribute
definition XML. This tag is to indicate that this attribute cannot be updated
from outside by attribute update chip-op. If the value of this attribute is
DISABLED or PERMISSIVE and allow attribute override bit-7 is set,
then SBE allows to update this attribute. But, will return the RC
SBE_ATTRIBUTE_RC_DENIED_UPDATE.

Change-Id: I2698ba792dbf19527cb0b8bcf346dee8e99368f7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb/+/165009
Reviewed-by: Akhilesh S <akhis023@in.ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/165015
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>