Enable generateSbeErrors for async SBE FFDC
This commit removes the stub interface for generateSbeErrors and enables
the function that handles SBE FFDC packages. This required numerous
small changes where the prior interface was called.

As a result, when async FFDC is requested via the genFifoSBEFFDCErrls
function the data will be returned to the caller.

Change-Id: I016434099473b2b03aadeb8677336d2f2d572fef
JIRA: PFHB-551
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/165028
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>