HTMGT: Use functional DIMM count calculating throttles
HTMGT determines how much power is available for each DIMM based on the
number of DIMMs in the system. Originally we were were using the number
of present DIMMs.
This is a workaround for RTC 596871 to correct the calculated DIMM count
when 32GB DDR5 DDIMMs are installed. The DIMMs only have a single port
vs other DDR5 DDIMMs that have 2 ports.

Change-Id: I157ae7f5a7fdda3d8f8044f9e105c2ce1348ad83
RTC: 596871
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/165130
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sheldon Bailey <baileysh@us.ibm.com>
Reviewed-by: Michael D Pardeik <pardeik@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martha Broyles <mbroyles@us.ibm.com>