DDR5 PMIC Health Check Enablement
Enable a periodic Hostboot Runtime collection of DDR5 Health Check
data (set to run every 6 hours).

When the DDR5 4U DDIMMs are healthy (determined by conditions
in the HWP's) then a response payload of a single byte will
be returned to Hostboot.

If a HWP returns a payload response size greater than one byte,
then Hostboot will log an error log with the payload returned
from the HWP.

This is to prevent excessive logging when PMIC's degrade.
PMIC's can transition from healthy to unhealthly, but yet
return again to a healthy state.  The minimal logging will
allow analysis to be done on the data.

This design also makes the interface more flexible for future
modifications in the HWP's to own the determination of when
and what data should be returned for logging.

JIRA: PFHB-493
Change-Id: I0a121e58cef06a4622c7ddd01e46e6f3a3037d73
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/165568
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>